renew project setup and refactoring preparations
moving to the Android Studio IDE, with this I had some issues with
ofcourse Gradle and other configurations settings. I have fixed these
issues and I started a preparations for refactoring the code for
instance for the positioning of the buildings. I have added a warehouse
building with placeholder graphics