drag road selection continuation
The drag functionality now works and just clicking works as well.
Removed a annoying but where it would register a single frame when the
build begins this is solved with an ugly boolean check. It also keeps
track of how many roads are in your inventory and doesn't allow more
than that to be build.