added region functionality
You now don't own any region on the Island, but you will when you build
a Warehouse, the region you will gain will be visible with the blue
region range you will see. When the Warehouse is build you have more
region and now you are able to build a farm or woodcutter, but only on
the region that you own.