warehouse quick fix
Made sure that you can, once again build the wareHouse and that it
rotates when it is placed in a spot that you can build. However there
are still issues with how it looks. The texture is not placed correctly
and it always shows all 9 tiles instead of just the 6 that matter.