simple camara start position update
I never changed the camera position. Since the (0, 0) mark is the left
most point of the map and the position of the camera was (viewwidht/2,
viewheight/2) it would always start a bit high and to the left. I
changed it to be (2000, 0) because the y = 0 will be the middle of the
map vertically and 2000 is about the middle horizontally