Test with a much much bigger map
I made the csv file that manages the map bigger. Now it is about the
size that the final game map should be able to be. However I noticed
that it gives me a 20 fps when debugging with desktop which leaves me to
believe that it will be about 1 fps on mobile. I'm thinking of
solutions.